+++ src/parser/enrichers/spell/FireShield.mjs
@@ -7,21 +7,82 @@ export default class FireShield extends DDBEnricherData {
     return "utility";
   }
 
-  get effects() {
+  get activity() {
+    return {
+      name: "Chill Shield",
+    };
+  }
+
+  get additionalActivities() {
     return [
       {
-        name: "Cold Shield",
-        changes: [
-          DDBEnricherData.ChangeHelper.unsignedAddChange("fire", 0, "system.traits.dr.value"),
-        ],
+        duplicate: true,
+        overrides: {
+          name: "Warm Shield",
+        },
       },
       {
-        name: "Warm Shield",
-        changes: [
-          DDBEnricherData.ChangeHelper.unsignedAddChange("cold", 0, "system.traits.dr.value"),
-        ],
+        constructor: {
+          name: "On Hit Damage",
+          type: "damage",
+        },
+        build: {
+          generateTarget: true,
+          generateDamage: true,
+          generateConsumption: false,
+          noSpellslot: true,
+          noeffect: true,
+        },
+        overrides: {
+          activationType: "special",
+          activationCondition: "A creature within 5 feet of you hits you with a melee attack roll",
+          data: {
+            damage: {
+              parts: [DDBEnricherData.basicDamagePart({
+                number: 2,
+                denomination: 8,
+                types: ["fire", "cold"],
+              })],
+            },
+          },
+        },
       },
     ];
+  }
+
+  get effects() {
+    return [
+      { name: "Chill Shield", damageType: "fire" },
+      { name: "Warm Shield", damageType: "cold" },
+    ].map((data) => {
+      return {
+        name: data.name,
+        activityMatch: data.name,
+        changes: [
+          DDBEnricherData.ChangeHelper.unsignedAddChange(data.damageType, 0, "system.traits.dr.value"),
+        ],
+        options: {
+          durationSeconds: 600,
+        },
+        onUseMacroChanges: [
+          { macroPass: "isDamaged", macroType: "spell", macroName: "fireShield.js", document: this.data },
+        ],
+        data: {
+          flags: {
+            dae: {
+              selfTargetAlways: true,
+            },
+          },
+        },
+      };
+    });
+  }
+
+  get itemMacro() {
+    return {
+      type: "spell",
+      name: "fireShield.js",
+    };
   }
 
 }