+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# Next Up
+
+- Ice Knife, Fire Shield Midi Automation
+
 # 6.0.62
 
 - When using CPR beta, allow application of CPR effects on character import.
-- Green Flame Blade and Booming Blade midi-auomations for v4.
+- Green Flame Blade and Booming Blade midi-automations for v4.
 
 # 6.0.61